feat(Board): 폴더 목록 구조 변경
