+++ src/main/java/com/finalproj/orbitflow/board/boardPermission/dto/BoardPermissionResDto.java
@@ -1,54 +1,14 @@
-package com.finalproj.orbitflow.board.dto.admin;
+package com.finalproj.orbitflow.board.boardPermission.dto;
 
-import com.finalproj.orbitflow.board.entity.BoardCategory;
-import com.finalproj.orbitflow.board.entity.BoardPermission;
+import com.finalproj.orbitflow.board.boardPermission.entity.BoardPermission;
 import com.finalproj.orbitflow.hr.employee.entity.Employee;
-import lombok.*;
+import lombok.Builder;
+import lombok.Getter;
 
-import java.time.Instant;
 import java.util.List;
 import java.util.stream.Collectors;
 
-public class AdminBoardResDto {
-
-    /**
-     * 게시판(카테고리) 목록 및 단건 조회 응답 DTO
-     */
-    @Data
-    @NoArgsConstructor
-    @AllArgsConstructor
-    @Builder
-    public static class Category {
-
-        private Long id; // 게시판 고유 ID
-        private Long companyId; // 회사 ID
-        private Long organizationId; // 조직 ID (null이면 일반 게시판)
-        private String boardName; // 게시판 이름
-        private String boardType; // 게시판 유형
-        private boolean isActivated; // 사용 여부
-        private boolean commentActivated; // 댓글 기능 활성화 여부
-        private Instant createdAt;
-        private Instant updatedAt;
-        private Instant deletedAt; // 삭제 일시
-
-        public static Category from(BoardCategory category) {
-            return Category.builder()
-                    .id(category.getId())
-                    // Company ID 추출
-                    .companyId(category.getCompany().getId())
-                    // Organization ID 추출 (null 체크)
-                    .organizationId(category.getOrganization() != null ? category.getOrganization().getId() : null)
-                    .boardName(category.getBoardName())
-                    .boardType(category.getBoardType())
-                    .isActivated(category.isActivated())
-                    .commentActivated(category.isCommentActivated())
-                    // BaseEntity에서 LocalDateTime을 사용한다고 가정
-                    .createdAt(category.getCreatedAt())
-                    .updatedAt(category.getUpdatedAt())
-                    .deletedAt(category.getDeletedAt())
-                    .build();
-        }
-    }
+public class BoardPermissionResDto {
 
     /**
      * 특정 게시판에 권한이 부여된 직원 목록 응답 DTO
